Refactor: JwtUtil 인터페이스로 수정
